CursorWorkplace/News Scraping/requirements.txt
# News Sentiment Scraper Dependencies

# Core scraping libraries
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
selenium==4.15.2
playwright==1.40.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# Sentiment analysis
textblob==0.17.1
nltk==3.8.1
vaderSentiment==3.3.2

# Database
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.0
sqlalchemy==1.4.54

# Scheduling and automation
apscheduler==3.10.4

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0

# Logging and monitoring
loguru==0.7.2

# Data visualization (optional)
plotly==5.17.0
streamlit==1.28.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0 